requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
pyttsx3
opencv-python
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.0
